A couple things
A. I changed the main app to be a viewport system, where each viewport interacts with stores that exist in the webapp that kind of just remain there.
B. Moved all the old html and javascript accordingly to a seperate file for the viewport.
C. Made text brighter, and added Cinzel for a more D&D style font.